wip: handle inbound ws message in a service to reuse the same logic with Blade
